Switch code from file to file
